modified:   ExternalAnalysisToMPAS.csh
	modified:   GetGFSAnalysisFromFTP.csh
	modified:   GetGFSAnalysisFromRDA.csh
	modified:   GetObs.csh
	modified:   LinkExternalAnalysis.csh
	modified:   ObsToIODA.csh
	modified:   UngribExternalAnalysis.csh
	new file:   config/jedi/ObsPlugs/variational/bias/abi_g16.yaml
	new file:   config/jedi/ObsPlugs/variational/bias/ahi_himawari8.yaml
	new file:   config/jedi/ObsPlugs/variational/filtersWithBias/abi_g16.yaml
	new file:   config/jedi/ObsPlugs/variational/filtersWithBias/ahi_himawari8.yaml
	modified:   config/observations.csh
	modified:   drive.csh
	modified:   include/criticalpath.rc
	modified:   include/tasks.rc
	modified:   include/verifymodel.rc
	modified:   include/verifyobs.rc
	modified:   run.csh
	deleted:    runs/ABI120km3denvar.yaml
	new file:   runs/ABIAHI120km3denvar.yaml
	renamed:    scenarios/ABI120km3denvar.yaml -> scenarios/ABIAHI120km3denvar.yaml
	modified:   scenarios/RealTime.yaml
	modified:   scenarios/base/observations.yaml
	modified:   scenarios/base/variational.yaml

+++ GetObs.csh
@@ -2,6 +2,19 @@
 # Get observations for a cold start experiment
 # from the NCEP FTP BUFR/PrepBUFR files or CISL RDA archived NCEP BUFR files
 
+# Process arguments
+# =================
+## args
+# ArgDT: int, valid time offset beyond CYLC_TASK_CYCLE_POINT in hours
+set ArgDT = "$1"
+
+set test = `echo $ArgDT | grep '^[0-9]*$'`
+set isNotInt = ($status)
+if ( $isNotInt ) then
+  echo "ERROR in $0 : ArgDT must be an integer, not $ArgDT"
+  exit 1
+endif
+
 date
 
 # Setup environment
@@ -10,12 +23,13 @@ source config/workflow.csh
 source config/observations.csh
 source config/experiment.csh
 source config/builds.csh
+source config/tools.csh
 set yyyymmdd = `echo ${CYLC_TASK_CYCLE_POINT} | cut -c 1-8`
-set ccyy = `echo ${CYLC_TASK_CYCLE_POINT} | cut -c1-4`
-set mmdd = `echo ${CYLC_TASK_CYCLE_POINT} | cut -c5-8`
+set ccyy = `echo ${CYLC_TASK_CYCLE_POINT} | cut -c 1-4`
+set mmdd = `echo ${CYLC_TASK_CYCLE_POINT} | cut -c 5-8`
 set hh = `echo ${CYLC_TASK_CYCLE_POINT} | cut -c 10-11`
 set thisCycleDate = ${yyyymmdd}${hh}
-set thisValidDate = ${thisCycleDate}
+set thisValidDate = `$advanceCYMDH ${thisCycleDate} ${ArgDT}`
 source ./getCycleVars.csh
 
 # templated work directory
@@ -49,10 +63,16 @@ foreach inst ( ${convertToIODAObservations} )
        endif
     # for prepbufr observations
     else if ( ${inst} == prepbufr ) then
-       setenv THIS_FILE prepbufr.gdas.${ccyy}${mmdd}.t${hh}z.nr.48h
+       setenv THIS_FILE prepbufr.gdas.${ccyy}${mmdd}${hh}.nr
        if ( ! -e ${THIS_FILE}) then
-          echo "Source file: ${PrepBUFRDirectory}/prep48h/${ccyy}/${THIS_FILE}"
-          cp -p ${PrepBUFRDirectory}/prep48h/${ccyy}/${THIS_FILE} .
+          if ( -e ${PrepBUFRDirectory}/prepnr/${ccyy}/${THIS_FILE} ) then
+             echo "Source file: ${PrepBUFRDirectory}/prepnr/${ccyy}/${THIS_FILE}"
+             cp -p ${PrepBUFRDirectory}/prepnr/${ccyy}/${THIS_FILE} .
+          else
+             setenv THIS_FILE prepbufr.gdas.${ccyy}${mmdd}.t${hh}z.nr.48h
+             echo "Source file: ${PrepBUFRDirectory}/prep48h/${ccyy}/${THIS_FILE}"
+             cp -p ${PrepBUFRDirectory}/prep48h/${ccyy}/${THIS_FILE} .
+          endif
        endif
        # use obs errors embedded in prepbufr file
        if ( -e obs_errtable ) then
